chore: transform searchbar to typescript
